Nemo-diarization/requirements.txt
# Nemo-Diarization Requirements
# These packages should already be installed in your ClearCast environment

torch>=2.3.1
torchaudio>=2.3.1
speechbrain>=1.0.0
resemblyzer>=0.1.4
openai-whisper>=20240930
numpy>=1.26.0,<2.0
scikit-learn>=1.7.0
scipy>=1.16.0
